requirements.txt
# Core dependencies
cryptography==42.0.5
pydantic==2.6.1
ollama==0.2.1
requests==2.31.0

# CLI and utilities
click==8.1.7
rich==13.7.1
python-dotenv==1.0.0

# Testing
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-xdist==3.5.0
responses==0.24.1

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.2

# Optional: For advanced caching
cachetools==5.3.3
